feat : Commands and Queries of Bookings and Homes implemented using Redis and repository implemented migration done

+++ src/01.Core/HouseRent.Core.ApplicationServices/Bookings/Commands/ReserveBookingCommand.cs
@@ -2,7 +2,7 @@
 
 namespace HouseRent.Core.ApplicationServices.Bookings.Commands;
 public record ReserveBookingCommand(
-    int HomeId,
-    int UserId,
+    long HomeId,
+    long UserId,
     DateOnly StartDate,
-    DateOnly EndDate) : ICommand<int>;
+    DateOnly EndDate) : ICommand<long>;
