+++ src/01.Core/HouseRent.Core.Domain/Homes/Events/HomeCreated.cs
@@ -2,4 +2,4 @@
 
 namespace HouseRent.Core.Domain.Homes.Events;
 
-public record HomeCreated(int id) : IDomainEvent;
+public record HomeCreated(long id) : IDomainEvent;
